pyproject.toml
[tool.poetry]
name = "dbus-scrobbler"
version = "1.0.0"
description = ""
authors = ["spezifisch <spezifisch@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = ">=3.5"
legacy-scrobbler = {git = "https://github.com/jlieth/legacy-scrobbler"}
click = "^7.0"
pygobject = "^3.32"
dbus-python = "^1.2"
mpris2 = "^1.0"
pyyaml = "^5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
scrobbler = "dbus_scrobbler.scrobbler:run"

